requirements.txt
asgiref==3.8.1
azure-core==1.34.0
azure-data-tables==12.7.0
azure-identity==1.23.0
azure-storage-blob==12.25.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
Django==5.2.3
django-cors-headers==4.7.0
django-environ==0.12.0
django-ratelimit==4.1.0
geopandas==1.1.0
gunicorn==23.0.0
idna==3.10
isodate==0.7.2
msal==1.32.3
msal-extensions==1.3.1
numpy==2.3.0
packaging==25.0
pandas==2.3.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
pyogrio==0.11.0
pyproj==3.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
shapely==2.1.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
